+++ comon_logger_dio/lib/comon_logger_dio.dart
@@ -1,3 +1,4 @@
 library comon_logger_dio;
 
 export 'src/comon_dio_interceptor.dart';
+export 'src/http_console_log_formatter.dart';