+++ comon_logger_navigation_flutter/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.1
+
+- Maintenance release for workspace version alignment.
+
 ## 0.1.0
 
 - Initial release of comon_logger_navigation_flutter.